requirements.txt
tweepy==4.11.0
python-dotenv==1.0.0
pyqt5==5.15.9
